Resolve http request query paremeters error
Before we tried passing in query parameters inside the query() function, but this resulted in 'TypeError: Right-hand side of 'instanceof' is not an object'.

Didn't really resolve the error, but found that passing the query parameters as a second argument into the get function worked. See 'Specifying request body' here https://bestofjs.org/projects/nock